rename: 모달 폴더 구조 변경
