+++ src/components/Investment/InvestmentPatch.jsx
@@ -2,10 +2,10 @@ import styles from "./InvestmentPatch.module.css";
 import X from "../../assets/ic_x.svg";
 import visibilityOff from "../../assets/btn_visibility_on.svg";
 import visibilityOn from "../../assets/btn_visibility_off.svg";
-import ModalContainer from "../Modal/ModalContainer";
 import { useState, useRef, useEffect } from "react";
 import InvestmentUpdate from "./InvestmentUpdate";
-import ConfirmModal from "../Modal/ConfirmModal";
+import ModalContainer from "../Modal/ModalContainer/ModalContainer";
+import ConfirmModal from "../Modal/ConfirmModal/ConfirmModal";
 
 export default function InvestmentPatch({ onClose, mockInvestor, startup }) {
   const { password: storedPassword } = mockInvestor || {};
